ai-service/requirements.txt
flask
flask-cors
ultralytics
opencv-python
numpy
requests
python-dotenv